% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yesno.R
\name{yesno}
\alias{yesno}
\title{Yes No}
\usage{
yesno(question)
}
\arguments{
\item{question}{A string of the question to ask.}
}
\value{
A flag indicating whether the user answered yes or no.
}
\description{
Asks a custom yes-no question with variable responses.
}

