% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yesno2.R
\name{yesno2}
\alias{yesno2}
\title{Yes No with Two Options}
\usage{
yesno2(...)
}
\arguments{
\item{...}{Objects to paste and then output to produce the question.}
}
\value{
A flag indicating whether the user answered yes or no.
}
\description{
Asks a custom yes-no question with 'Yes' or 'No' (in that order) as the only responses.
Returns a flag indicating which response the user choose.
It is designed to be used in situations where a user needs to choose one of two options both of which have consequences.
}
\details{
The objects are first pasted without separators
and collapsed using \code{\link{paste0}(..., collapse = "")}
before being output using \code{\link{cat}}.

The order and phrasing is fixed so that users can quickly type their
preferred response to the question.

Selection of a 0 (to exit) is also considered to be No.
}
\examples{
\dontrun{
yesno2("Do you like ", R.Version()$nickname ,"?")
}
}
\seealso{
\code{\link{yesno}}
}
